% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCmed.R
\docType{package}
\name{CCmed}
\alias{CCmed}
\alias{CCmed-package}
\title{CCmed: A package to conduct cross-condition mediation analysis.}
\description{
The CCmed package implements two versions of the CCmed algorithm:
CCmed_gene and CCmed_gwas.
}
\section{CCmed functions}{

To estimate the probability of gene-level trans-associations mediated
by cis-associations, see \code{\link{CCmed_gene}}.
To estimate the probability of trans-associations for GWAS SNPs mediated
by cis-associations, see \code{\link{CCmed_gwas}}.
}

