% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CCmed_gene}
\alias{CCmed_gene}
\title{Estimate cross-condition mediation probability at the gene-level.}
\usage{
CCmed_gene(cis_res, condcorr_res, cisID, condcorrID, ncond)
}
\arguments{
\item{cis_res}{list object returned by running Primo on gene-level cis-association statistics.}

\item{condcorr_res}{list object returned by running Primo on conditional correlation statistics.}

\item{cisID}{vector of gene names corresponding that match the identifiers of of \code{cis_res} (e.g. rows in
list element \code{post_prob}).}

\item{condcorrID}{data.frame with two columns ("cis", "trans") that match the identifiers
of \code{condcorr_res} (i.e. rows in list element \code{post_prob}).}

\item{ncond}{integer of the number of conditions required for mediation.}
}
\value{
A data.frame with the following columns:
\tabular{ll}{
\code{cis} \tab cis-gene identifier\cr
\code{trans} \tab trans-gene identifier\cr
\code{ncond} \tab number of conditions required for mediation\cr
\code{prob_cis_all} \tab posterior probability of cis-association in all conditions\cr
\code{prob_condcorr_ncond} \tab posterior probability of conditional correlation
in at least \code{ncond} conditions\cr
\code{prob_med} \tab probability of mediation in at least \code{ncond} conditions
}

The main element of interest for inference is the column \code{prob_med}.
}
\description{
For each cis-trans gene pair, estimate the probability of mediation.
}
