% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CCmed_gwas}
\alias{CCmed_gwas}
\title{Estimate cross-condition mediation probability for a GWAS SNP.}
\usage{
CCmed_gwas(cis_res, condcorr_res, cisID, condcorrID, ncond)
}
\arguments{
\item{cis_res}{list object returned by running Primo on GWAS SNP cis-association statistics.}

\item{condcorr_res}{list object returned by running Primo on conditional correlation statistics.}

\item{cisID}{data.frame with two columns ("variant_id", "cis") that match the identifiers of
\code{cis_res} (i.e. rows in list element \code{post_prob}).}

\item{condcorrID}{data.frame with two columns ("cis", "trans") that match the identifiers
of \code{condcorr_res} (i.e. rows in list element \code{post_prob}).}

\item{ncond}{integer of the number of conditions required for mediation.}
}
\value{
data.frame with the following columns:
\tabular{ll}{
\code{variant_id} \tab GWAS SNP identifier\cr
\code{cis} \tab cis-gene identifier\cr
\code{trans} \tab trans-gene identifier\cr
\code{ncond} \tab number of conditions required for mediation\cr
\code{prob_med} \tab probability of mediation in at least \code{ncond} conditions
}

The main element of interest for inference is the column \code{prob_med}.
}
\description{
For each SNP-cis-trans gene trio, estimate the probability of mediation.
Used to identify trans-associations of GWAS SNPs.
.
}
